+++ module-11/src/main/java/com/epam/kabaldin/model/UserAccount.java
@@ -1,9 +1,11 @@
 package com.epam.kabaldin.model;
 
+import java.math.BigDecimal;
+
 public interface UserAccount {
 
     Long getId();
     void setId(Long id);
-    Long getPrepaidMoney();
-    void setPrepaidMoney(Long prepaidMoney);
+    BigDecimal getPrepaidMoney();
+    void setPrepaidMoney(BigDecimal prepaidMoney);
 }
