+++ module-11/src/main/java/com/epam/kabaldin/service/UserAccountService.java
@@ -2,11 +2,10 @@ package com.epam.kabaldin.service;
 
 import com.epam.kabaldin.model.UserAccount;
 
-import java.util.Optional;
+import java.math.BigDecimal;
 
 public interface UserAccountService {
     UserAccount getUserAccountById(Long id);
     boolean updateUserAccount(UserAccount userAccount);
-
-    void refillAccount(Long userId, Long amount);
+    void refillAccount(Long userId, BigDecimal amount);
 }
